refactor: move surface playback scenarios
