+++ tests/coding/test_native_tui_playback_runner.py
@@ -12,6 +12,7 @@ from loushang.coding.ui.playback_runner import (
     run_playback_scenarios,
 )
 from loushang.coding.ui.playback_scenarios.command import COMMAND_ROUTING_SCENARIOS
+from loushang.coding.ui.playback_scenarios.surface import SURFACE_SCENARIOS
 from loushang.coding.ui.playback_suite import NativePlaybackSuite as SuiteFromModule
 
 
@@ -38,6 +39,23 @@ def test_native_tui_playback_command_scenarios_live_in_command_module() -> None:
         "session-command-error",
         "unknown-slash-prompt",
         "non-executable-session-command",
+    ]
+
+
+def test_native_tui_playback_surface_scenarios_live_in_surface_module() -> None:
+    assert [scenario.name for scenario in SURFACE_SCENARIOS] == [
+        "status-surface",
+        "statusline-command",
+        "command-palette-select",
+        "command-palette-session-command",
+        "commands-info-surface",
+        "commands-info-session-command",
+        "settings-search",
+        "model-select",
+        "model-select-search",
+        "approval-surface",
+        "approval-reject-surface",
+        "dialog-surface",
     ]
